RoomCardをcomponentディレクトリに移動し、名前をRoomPreviewCardに変更
